Creates views and adds form handling
Chose not to use ModelForms. This was for two reasons: First, I didn't
want the models to preclude javascript webcomponents in the form.
Second, I don't know how they work yet! In retrospect, throwing away a
considerable amount of free validation (all the way back to the db) is
not worth the reducing the *possibility* of using client-side widgets.

For this project since learning django is primary, ModelForms would have
been a better choice.